Moved selection.service.ts and its spec file to services folder and updated imports to said files [#158037410]
