+++ projects/revaturecloud/selection/src/lib/selection.component.ts
@@ -5,7 +5,7 @@
 */
 
 import { Component, OnInit } from '@angular/core';
-import { SelectionService } from './selection.service';
+import { SelectionService } from './services/selection.service';
 import { Room } from './models/room.model';
 import { Observable } from 'rxjs';
 import { RoomAssociation } from './models/roomAssociation.model';
