+++ projects/revaturecloud/selection/src/public_api.ts
@@ -2,6 +2,6 @@
  * Public API Surface of selection
  */
 
-export * from './lib/selection.service';
+export * from './lib/services/selection.service';
 export * from './lib/selection.component';
 export * from './lib/selection.module';
